fix: complete security hardening and edge-case closure
This finalizes strict parser and auth hardening, memory-lifetime safety, checked wrappers, and crypto-boundary reliability so core trust-boundary behavior is deterministic and safer-by-default.

+++ src/nip70_protected.zig
@@ -4,10 +4,10 @@ const nip01_event = @import("nip01_event.zig");
 pub const ProtectedError = error{ ProtectedAuthRequired, ProtectedPubkeyMismatch };
 
 pub fn event_has_protected_tag(event: *const nip01_event.Event) bool {
-    std.debug.assert(event.tags.len <= std.math.maxInt(u16));
+    std.debug.assert(event.tags.len <= std.math.maxInt(usize));
     std.debug.assert(@intFromPtr(event) != 0);
 
-    var index: u16 = 0;
+    var index: usize = 0;
     while (index < event.tags.len) : (index += 1) {
         const tag = event.tags[index];
         if (tag.items.len != 1) {
